docs(Select): one example per story
Example and TriggerSlots each rendered two side-by-side demos with
their own headings. Split them so each story shows one thing.

Drops the States story — its disabled option is already visible in
Example — and the prefix/suffix half of TriggerSlots, which OptionSlot
already covers.
